+++ ConfigManager.cs
@@ -16,6 +16,7 @@ public class ConfigManager {
     public static ColorField AccentColor;
     public static BoolField EnableTrackers;
     public static IntField TrackerThreshold;
+    public static Dictionary<string, BoolField> ForceTrackEnemies = new Dictionary<string, BoolField>();
     public static BoolField TrackerIgnorePuppets;
     public static BoolField TrackerShowEnemyNames;
     public static ColorField TrackerColor;
@@ -31,7 +32,7 @@ public class ConfigManager {
     public static FloatField TrackerScale;
     public static FloatSliderField TrackerMarkOpacity;
 
-    public static Dictionary<string, BoolField> ForceTrackEnemies = new Dictionary<string, BoolField>();
+    public static StringField LastVersion;
 
     public static void Initialize() {
     }
@@ -41,7 +42,7 @@ public class ConfigManager {
     }
 
     private static void CreateTrackerForcedEnemies() {
-        ConfigPanel forceTrackerPanel = new ConfigPanel(config.rootPanel, "Always show tracker for specific enemies", "forcedTrackEnemies");
+        ConfigPanel forceTrackerPanel = new ConfigPanel(config.rootPanel, "Always track enemy types...", "forcedTrackEnemies");
         new ConfigHeader(config.rootPanel, "", 10);
         new ConfigHeader(forceTrackerPanel, "You can, for example, force Powers and Mindflayers to always have an indicator shown. FUCK POWERS FUCK POWERS FUCK POWERS I HATE POWERS", 12, TextAlignmentOptions.Left);
         string[] names = Enum.GetNames(typeof(EnemyType));
@@ -58,21 +59,23 @@ public class ConfigManager {
         if (File.Exists(iconPath)) config.SetIconWithURL(iconPath);
 
         new ConfigHeader(config.rootPanel, "", 10);
-        new ConfigHeader(config.rootPanel, "-- DEATH CROSSES --", 22);
+        new ConfigHeader(config.rootPanel, "-- <color=#db1e39>DEATH CROSSES</color> --", 22);
         Enable = new BoolField(config.rootPanel, "Enable death crosses", "enable", true);
         AccentColor = new ColorField(config.rootPanel, "Cross accent color", "accentColor", new Color(219f / 255f, 30f / 255f, 57f / 255f));
 
         new ConfigHeader(config.rootPanel, "", 10);
-        new ConfigHeader(config.rootPanel, "-- ENEMY TRACKERS --", 22);
-        EnableTrackers = new BoolField(config.rootPanel, "Enable trackers", "enableTrackers", false);
+        new ConfigHeader(config.rootPanel, "-- <color=#f1b613>ENEMY TRACKERS</color> --", 22);
+        new ConfigHeader(config.rootPanel, "Shows pointers for enemy locations even when they're out of view", 12);
+        EnableTrackers = new BoolField(config.rootPanel, "Enable tracker icons", "enableTrackers", false);
         TrackerThreshold = new IntField(config.rootPanel, "Reveal last remaining x enemies", "trackerThreshold", 5);
         CreateTrackerForcedEnemies();
         TrackerIgnorePuppets = new BoolField(config.rootPanel, "Ignore puppets (blood bois)", "trackerIgnorePuppets", true);
         TrackerShowEnemyNames = new BoolField(config.rootPanel, "Show enemy names", "trackerShowEnemyNames", true);
-        TrackerColor = new ColorField(config.rootPanel, "Enemy tracker mark color", "trackerColor", new Color(241f / 255f, 182f / 255f, 19f / 255f));
+        TrackerColor = new ColorField(config.rootPanel, "Tracker accent color", "trackerColor", new Color(241f / 255f, 182f / 255f, 19f / 255f));
 
         new ConfigHeader(config.rootPanel, "", 10);
-        new ConfigHeader(config.rootPanel, "-- ADVANCED --", 20);
+        new ConfigHeader(config.rootPanel, "-- <color=#ff7f27>ADVANCED</color> --", 22);
+        new ConfigHeader(config.rootPanel, "// <color=#db1e39>Death crosses</color>", 16, TextAlignmentOptions.Left);
         StartScale = new FloatField(config.rootPanel, "Cross starting scale", "startScale", 2f);
         EndScale = new FloatField(config.rootPanel, "Cross ending scale", "endScale", 0.35f);
         EnemyHealthThreshold = new FloatField(config.rootPanel, "Cross enemy health threshold", "enemyHealthThreshold", 0);
@@ -81,9 +84,13 @@ public class ConfigManager {
         ScalingDelay = new FloatField(config.rootPanel, "Scaling delay", "scalingDelay", 0.000f);
         ScalingDuration = new FloatField(config.rootPanel, "Scaling duration", "scalingDuration", 0.1333f);
         VisibleDuration = new FloatField(config.rootPanel, "Visible duration", "visibleDuration", 0.5333f);
-
+        new ConfigHeader(config.rootPanel, "// <color=#f1b613>Enemy trackers</color>", 16, TextAlignmentOptions.Left);
         TrackerScale = new FloatField(config.rootPanel, "Tracker mark scale", "trackerScale", 1f);
         TrackerMarkOpacity = new FloatSliderField(config.rootPanel, "Tracker mark opacity", "trackerMarkOpacity",
             new Tuple<float, float>(0f, 1f), 1f);
+
+        // Internal
+        LastVersion = new StringField(config.rootPanel, "Last version", "lastVersion", "0.0.0");
+        LastVersion.hidden = true;
     }
 }