+++ package/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.1
 
+- Made cross durations customizable
 - Internal: create crosses when enemies die instead of when they spawn. Should help performance a bit.
 
 ## 1.0.0